+++ implementation/js/src/reusable/widgets/datagrid-lite/index.js
@@ -74,9 +74,9 @@
 
 				//2. rebuild header cells - let it rerender with new column array
 				_.each(this._options.columns, function(column){
-					column.header = column.header || 'string',
-					column.cell = column.cell || column.header || 'string',
-					column.label = column.label || _.string.titleize(column.name)
+					column.header = column.header || 'string';
+					column.cell = column.cell || column.header || 'string';
+					column.label = column.label || _.string.titleize(column.name);
 				});				
 				this.header.currentView.trigger('view:render-data', this._options.columns);
 
@@ -107,15 +107,19 @@
 			type: 'CollectionView',
 			itemView: 'dynamic',
 			tagName: 'tr',
+			initialize: function(options){
+				this.grid = options.body.parentCt; //give each row the grid view ref.
+			},
 			//buildItemView - select proper cell
 			buildItemView: function(item, ItemViewType, itemViewOptions){
 				return app.widget(_.string.classify([item.get('cell'), 'cell'].join('-')), {
-					model: item,
 					tagName: 'td',
+					model: item,
+
 					row: this //link each cell with the row. (use/link it in cell's init())
 				});
 			}			
-		})		
+		});
 
 		var Body = app.view({
 			type: 'CollectionView',
@@ -126,12 +130,13 @@
 						return _.extend({
 							value: selectn(column.name || '', model.attributes),
 							index: index
-						}, column)
-					}, this))
-				}
-			}
-		})
+						}, column);
+					}, this)),
 
+					body: this //passing body to row view
+				};
+			}
+		});
 		return UI;
 
 	});